testing committer pain -pull out some into their own -log more @ debug -fixing inadequate tes setup
+ -Ddebug option to halt scalatest JVM

+++ cloud-examples/src/test/scala/org/apache/spark/sql/sources/AbstractCloudRelationTest.scala
@@ -20,17 +20,15 @@ package org.apache.spark.sql.sources
 import java.io.File
 
 import scala.concurrent.duration._
-import scala.util.Random
 
-import com.hortonworks.spark.cloud.CloudSuiteTrait
-import com.hortonworks.spark.cloud.s3.S3ACommitterConstants._
-import org.apache.hadoop.fs.Path
+import com.hortonworks.spark.cloud.{CloudSuiteTrait, ObjectStoreConfigurations}
+import org.apache.hadoop.fs.{FileStatus, Path}
 import org.scalatest.BeforeAndAfterAll
 import org.scalatest.concurrent.Eventually
 
 import org.apache.spark.SparkContext
 import org.apache.spark.sql._
-import org.apache.spark.sql.hive.test.{TestHive, TestHiveSingleton}
+import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.test.SQLTestUtils
 import org.apache.spark.sql.types._
 import org.apache.spark.util.Utils
@@ -43,7 +41,7 @@ import org.apache.spark.util.Utils
  */
 abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
   with Eventually
-  with TestHiveSingleton
+  with HiveSingletonTrait
   with CloudSuiteTrait with BeforeAndAfterAll {
 
   import spark.implicits._
@@ -53,6 +51,17 @@ abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
   protected override def beforeAll(): Unit = {
     super.beforeAll()
 //    COMMITTERS_BY_NAME(DIRECTORY).bind()
+
+
+    // validate the conf by asserting that the spark conf is bonded
+    // to the partitioned committer.
+    val sparkConf = spark.conf
+    assert(ObjectStoreConfigurations.PARQUET_COMMITTER_CLASS ===
+        sparkConf.get(SQLConf.PARQUET_OUTPUT_COMMITTER_CLASS.key),
+      s"wrong value of ${SQLConf.PARQUET_OUTPUT_COMMITTER_CLASS}")
+    assert(ObjectStoreConfigurations.PATH_OUTPUT_COMMITTER_NAME ===
+        sparkConf.get(SQLConf.FILE_COMMIT_PROTOCOL_CLASS.key),
+      s"wrong value of ${SQLConf.PARQUET_OUTPUT_COMMITTER_CLASS}")
   }
 
   protected override def afterAll(): Unit = {
@@ -93,37 +102,101 @@ abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
 
   lazy val testDF = spark.range(1, 3).map(i => (i, s"val_$i")).toDF("a", "b")
 
-  lazy val partitionedTestDF1 = (for {
+  lazy val partitionedTestDF1: DataFrame = (for {
     i <- 1 to 3
     p2 <- Seq("foo", "bar")
   } yield {
     (i, s"val_$i", 1, p2)
   }).toDF("a", "b", "p1", "p2")
 
-  lazy val partitionedTestDF2 = (for {
+  lazy val partitionedTestDF2: DataFrame = (for {
     i <- 1 to 3
     p2 <- Seq("foo", "bar")
   } yield {
     (i, s"val_$i", 2, p2)
   }).toDF("a", "b", "p1", "p2")
 
-  lazy val partitionedTestDF = partitionedTestDF1.union(partitionedTestDF2)
+  protected lazy val partitionedTestDF: DataFrame =
+    partitionedTestDF1.union(partitionedTestDF2)
+
+  /**
+   * Get
+   * @return
+   */
+  protected def defaultOutputValidator(): Option[Path => Unit] = {
+    Some(assertSuccessFileExists)
+  }
+
+  /**
+   * Verify that the _SUCCESS file exists in a directory.
+   * @param destDir destination dir
+   */
+  protected def assertSuccessFileExists(destDir: Path): Unit = {
+    resolveSuccessFile(destDir, true)
+  }
+
+  /**
+   * Get the success file
+   * @param destDir destination of the query
+   * @param requireS3ACommitter is an S3A committer required
+   * @return the status
+   */
+  protected def resolveSuccessFile(destDir: Path, requireS3ACommitter: Boolean): FileStatus = {
+    val fs = getFilesystem(destDir)
+    val success = new Path(destDir, "_SUCCESS")
+    val status = fs.getFileStatus(success)
+    if (status.getLen == 0) {
+      logInfo(s"Status file $success exists and is an empty files")
+      assert(!requireS3ACommitter,
+        s"The committer used to create file $success was not an S3A Committer")
+    } else {
+      logInfo(s"Status file $success is of size ${status.getLen}")
+    }
+    status
+  }
 
   /**
    * Generates a temporary path without creating the actual file/directory, pass
    * it to the function, then cleanup with a deleteQuietly().
+   * A validator function can be supplied to validate the data.
    *
-   * @todo Probably this method should be moved to a more general place
+   * @param name test name, used in path calculation
+   * @param validator optional validator function. The default is that returned by
+   *                  defaultOutputValidator.
+   * @param fn function to evaluate
    */
-  protected def withPath(name: String)(f: Path => Unit): Unit = {
-    val dir = path(name)
+  protected def withPath(name: String,
+    validator: Option[Path => Unit] = defaultOutputValidator())
+    (fn: Path => Unit): Path = {
+    val d = path(name)
+    filesystem.delete(d, true)
+    withDefinedPath(d, fn, validator)
+  }
+
+  /**
+   * Execute the function within the defined path, which is deleted afterwards.
+   * @param dir directory
+   * @param fn fun to invoke
+   * @param validator validator to run after the operation
+   * @param deleteAfter should the dir be deleted after?
+   * @return the directory
+   */
+  private def withDefinedPath(dir: Path,
+    fn: Path => Unit,
+    validator: Option[Path => Unit],
+    deleteAfter: Boolean = true): Path = {
     try {
-      f(dir)
+      fn(dir)
+      waitForCompletion()
+      validator.foreach(p => p.apply(dir))
     } finally {
       // wait for all tasks to finish before deleting files
       waitForCompletion()
-      deleteQuietly(dir)
+      if (deleteAfter) {
+        deleteQuietly(dir)
+      }
     }
+    dir
   }
 
   /**
@@ -142,31 +215,28 @@ abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
    * Creates a temporary directory, which is then passed to `f` and will be deleted after `f`
    * returns.
    */
-  protected def withTempPathDir(name: String)(f: Path => Unit): Unit = {
+  protected def withTempPathDir(
+    name: String,
+    validator: Option[Path => Unit] = defaultOutputValidator(),
+    deleteAfter: Boolean = true)
+    (fn: Path => Unit): Path = {
     val dir = path(name)
+    filesystem.delete(dir, true)
     filesystem.mkdirs(dir)
-    try {
-      f(dir)
-    } finally {
-      // wait for all tasks to finish before deleting files
-      waitForCompletion()
-      deleteQuietly(dir)
-    }
+    withDefinedPath(dir, fn, validator, deleteAfter)
   }
-
 
   /**
    * Generates a temporary path without creating the actual file/directory, then pass it to `f`.
-   * If
-   * a file/directory is created there by `f`, it will be deleted after `f` returns.
+   * If a file/directory is created there by `f`, it will be deleted after `f` returns.
    *
    * @todo Probably this method should be moved to a more general place
    */
-  protected override def withTempPath(f: File => Unit): Unit = {
+  protected override def withTempPath(fn: File => Unit): Unit = {
     val path = Utils.createTempDir()
     path.delete()
     try {
-      f(path)
+      fn(path)
     } finally {
       Utils.deleteRecursively(path)
     }
@@ -225,7 +295,7 @@ abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
     }
   }
 
-  private val supportedDataTypes = Seq(
+  protected val parquetDataTypes: Seq[DataType] = Seq(
     StringType, BinaryType,
     NullType, BooleanType,
     ByteType, ShortType, IntegerType, LongType,
@@ -239,7 +309,7 @@ abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
     new UDT.MyDenseVectorUDT()
   ).filter(supportsDataType)
 
-  private val orcSupportedDataTypes = Seq(
+  protected val orcSupportedDataTypes: Seq[DataType] = Seq(
     StringType, BinaryType,
     NullType, BooleanType,
     ByteType, ShortType, IntegerType, LongType,
@@ -253,60 +323,6 @@ abstract class AbstractCloudRelationTest extends QueryTest with SQLTestUtils
     new UDT.MyDenseVectorUDT()
   )
 
-  for (dataType <- supportedDataTypes) {
-    for (parquetDictionaryEncodingEnabled <- Seq(true, false)) {
-      ctest(
-        s"test all data types - $dataType with parquet.enable.dictionary = " +
-          s"$parquetDictionaryEncodingEnabled", "", false) {
-
-        val extraOptions = Map[String, String](
-          "parquet.enable.dictionary" ->
-            parquetDictionaryEncodingEnabled.toString
-        )
-
-        withTempPath { file =>
-          val path = file.toString
-
-          val dataGenerator = RandomDataGenerator.forType(
-            dataType = dataType,
-            nullable = true,
-            new Random(System.nanoTime())
-          ).getOrElse {
-            fail(s"Failed to create data generator for schema $dataType")
-          }
-
-          // Create a DF for the schema with random data. The index field is used to sort the
-          // DataFrame.  This is a workaround for SPARK-10591.
-          val schema = new StructType()
-            .add("index", IntegerType, nullable = false)
-            .add("col", dataType, nullable = true)
-          val rdd =
-            spark.sparkContext
-              .parallelize((1 to 10).map(i => Row(i, dataGenerator())))
-          val df = spark.createDataFrame(rdd, schema).orderBy("index")
-            .coalesce(1)
-
-          df.write
-            .mode("overwrite")
-            .format(dataSourceName)
-            .option("dataSchema", df.schema.json)
-            .options(extraOptions)
-            .save(path)
-
-          val loadedDF = spark
-            .read
-            .format(dataSourceName)
-            .option("dataSchema", df.schema.json)
-            .schema(df.schema)
-            .options(extraOptions)
-            .load(path)
-            .orderBy("index")
-
-          checkAnswer(loadedDF, df)
-        }
-      }
-    }
-  }
 
 
 }
